deploy-your-ai-application-in-production/codefiles/requirements.txt
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
openai==1.12.0
